+++ release-notes.md
@@ -1,6 +1,9 @@
 Version 1.3.1 (unreleased)
 --------------
 * Fixed zone ID calculation for non-uniform grids.  See PR #2.
+* Modified ZoneManager to automatically send "no-change" updates for
+    objects it is managing but didn't receive updates for.  See PR #5
+
 
 Version 1.3.0 (latest)
 --------------